Fix `perps/utils.ts` circular dependencies
