+++ src/features/perps/depositConfig.ts
@@ -4,7 +4,7 @@ import { time } from '@/framework/core/utils/time';
 import { createDepositConfig } from '@/systems/funding/config';
 
 import { HYPERCORE_PSEUDO_CHAIN_ID, HYPERLIQUID_USDC_ADDRESS, USDC_ICON_URL } from './constants';
-import { refetchHyperliquidBalance } from './utils';
+import { refetchHyperliquidBalance } from './stores/hyperliquidAccountStore';
 
 // ============ Perps Deposit Configuration ================================= //
