+++ src/features/perps/screens/PerpsSearchScreen.tsx
@@ -8,10 +8,10 @@ import { PerpMarketsList } from '@/features/perps/components/PerpMarketsList';
 import { FOOTER_HEIGHT, PERPS_BACKGROUND_DARK, PERPS_BACKGROUND_LIGHT } from '@/features/perps/constants';
 import { usePerpsAccentColorContext } from '@/features/perps/context/PerpsAccentColorContext';
 import { useMarketSortOrderLabels } from '@/features/perps/hooks/useMarketSortOrderLabels';
+import { navigateToNewPositionScreen, navigateToPerpDetailScreen } from '@/features/perps/navigation/marketNavigation';
 import { PerpsNavigation, usePerpsNavigationStore } from '@/features/perps/navigation/perpsNavigation';
 import { useFilteredHyperliquidMarkets, useHyperliquidMarketsStore } from '@/features/perps/stores/hyperliquidMarketsStore';
 import { type PerpMarket } from '@/features/perps/types';
-import { navigateToNewPositionScreen, navigateToPerpDetailScreen } from '@/features/perps/utils';
 import { time } from '@/framework/core/utils/time';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { useOnLeaveRoute } from '@/hooks/useOnLeaveRoute';
