+++ src/features/perps/screens/perp-detail-screen/SheetFooter.tsx
@@ -8,9 +8,9 @@ import { EasingGradient } from '@/components/easing-gradient/EasingGradient';
 import { Box, Text, useColorMode } from '@/design-system';
 import { useAddCashRoute } from '@/features/cash/navigation/useAddCashRoute';
 import { HyperliquidButton } from '@/features/perps/components/HyperliquidButton';
+import { navigateToNewPositionScreen } from '@/features/perps/navigation/marketNavigation';
 import { useHyperliquidAccountStore } from '@/features/perps/stores/hyperliquidAccountStore';
 import { type PerpMarket } from '@/features/perps/types';
-import { navigateToNewPositionScreen } from '@/features/perps/utils';
 import * as i18n from '@/languages';
 import Navigation from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
