+++ src/features/perps/screens/perps-account-screen/OpenPositionsSection.tsx
@@ -6,8 +6,8 @@ import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { Box, Stack, Text, TextShadow } from '@/design-system';
 import { PerpPositionCard } from '@/features/perps/components/PerpPositionCard';
 import { DOWN_ARROW, HYPERLIQUID_COLORS, UP_ARROW } from '@/features/perps/constants';
+import { navigateToPerpDetailScreen } from '@/features/perps/navigation/marketNavigation';
 import { usePerpsPositionsInfo } from '@/features/perps/stores/derived/usePerpsPositionsInfo';
-import { navigateToPerpDetailScreen } from '@/features/perps/utils';
 import * as i18n from '@/languages';
 import Navigation from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
